refactor: Extract recipe, fix styles
