+++ libs/novui/src/panda-preset.ts
@@ -24,6 +24,7 @@ import { SEMANTIC_SPACING_TOKENS } from './tokens/semanticSpacing.tokens';
 import { SEMANTIC_RADIUS_TOKENS } from './tokens/semanticRadius.tokens';
 import { LEGACY_OPACITY_TOKENS, OPACITY_TOKENS } from './tokens/opacity.tokens';
 import { SEMANTIC_OPACITY_TOKENS } from './tokens/semanticOpacity.tokens';
+import { BUTTON_RECIPE } from './recipes/button.recipe';
 
 /**
  * This defines all Novu tokens into a single preset to be used in our various apps (and design-system).
@@ -83,6 +84,7 @@ export const novuPandaPreset = definePreset({
       recipes: {
         text: TEXT_RECIPE,
         title: TITLE_RECIPE,
+        button: BUTTON_RECIPE,
       },
     },
   },
